% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_qc_phase1.2.R
\name{runQC_Phase1.2_report}
\alias{runQC_Phase1.2_report}
\title{Run QC}
\usage{
runQC_Phase1.2_report(
  file.nm1,
  dat.DailyCounts,
  dat.ClinicalCourse,
  dat.AgeSex,
  dat.DiagProcMed,
  dat.Labs,
  dat.RaceByLocalCode,
  dat.RaceBy4CECode,
  dat.LabCodes,
  icd.list,
  lab.range,
  site.nm
)
}
\arguments{
\item{file.nm1}{path and name of output pdf (string)}

\item{dat.DailyCounts}{DailyCounts}

\item{dat.ClinicalCourse}{ClinicalCourse}

\item{dat.AgeSex}{AgeSex}

\item{dat.DiagProcMed}{DiagProcMed}

\item{dat.Labs}{Labs}

\item{dat.RaceByLocalCode}{RaceByLocalCode}

\item{dat.RaceBy4CECode}{RaceBy4CECode}

\item{dat.LabCodes}{LabCodes}

\item{icd.list}{list of icd codes}

\item{lab.range}{range of normal lab values}

\item{site.nm}{site name (string)}
}
\value{
QC file
}
\description{
Run QC
}
